Agregar botón Cambiar contraseña

+++ extra-hours-amadeus/src/main/java/com/example/extra_hours_amadeus/repository/UsersRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UsersRepo extends JpaRepository<Users, Integer> {
+public interface UsersRepo extends JpaRepository<Users, Long> {
 
     Optional<Users> findByEmail(String email);
 }
